Moved board types to separate file
It will simplify using them on the frontend (boardClient) and the backend
